refactor: prompt コンテキストを evaluation にリネーム
merge-ready prompt サブコマンドの廃止により「prompt」という名前が
実態（PR マージ可否の評価）を表さなくなったため evaluation に変更する。
