+++ src/contexts.rs
@@ -3,7 +3,7 @@
 //! This module exposes three contexts, each separated by responsibility.
 //!
 //! - [`config`] - manages display and behavior settings
-//! - [`prompt`] - evaluates pull request merge readiness
+//! - [`evaluation`] - evaluates pull request merge readiness
 //! - [`daemon`] - provides low-latency responses via daemon/cache
 //!
 //! # Context relationship
@@ -11,10 +11,10 @@
 //! ```text
 //! merge-ready-prompt (lightweight bin)
 //!   -> daemon (serve quickly from daemon/cache)
-//!   -> prompt (fetch + evaluate merge readiness)
+//!   -> evaluation (fetch + evaluate merge readiness)
 //!   -> config (presentation and behavior settings)
 //! ```
 
 pub mod config;
 pub mod daemon;
-pub mod prompt;
+pub mod evaluation;
