+++ src/contexts/evaluation/application/branch_sync.rs
@@ -1,5 +1,5 @@
-use crate::contexts::prompt::domain::branch_sync::{BranchSync, BranchSyncRepository};
-use crate::contexts::prompt::domain::error::RepositoryError;
+use crate::contexts::evaluation::domain::branch_sync::{BranchSync, BranchSyncRepository};
+use crate::contexts::evaluation::domain::error::RepositoryError;
 
 /// ブランチ同期状態を取得する。失敗時は `Err` を返す（エラー表示は呼び出し元が担う）。
 pub fn fetch(repo: &impl BranchSyncRepository) -> Result<BranchSync, RepositoryError> {
