+++ src/contexts/evaluation/application/merge_ready.rs
@@ -1,5 +1,5 @@
 use super::errors::{ErrorLogger, ErrorPresenter};
-use crate::contexts::prompt::domain::merge_ready::{MergeReadiness, MergeReadinessRepository};
+use crate::contexts::evaluation::domain::merge_ready::{MergeReadiness, MergeReadinessRepository};
 
 /// マージ可否状態を取得する。失敗時は `None` を返してエラー出力する。
 pub fn fetch(
