+++ src/contexts/evaluation/domain/ci_checks/aggregate.rs
