+++ src/contexts/evaluation/domain/signal.rs
