--- src/contexts/prompt/interface/cli.rs
@@ -1,1 +1,0 @@
-
